Add latency in `AdmissionController::returnedResponse`
Summary:
Currently, it is not possible to measure latency of responses in the `AdmissionController`.
This diff introduce a slight change in API, now the `returnedResponse` method is called with a Clock::duration representing the latency of the response.

`ResponseChannel` has been modified to now measure its lifetime which translate to the processing time of a response. It passes that information to the `AdmissionController`.

Reviewed By: yfeldblum

Differential Revision: D13554695

fbshipit-source-id: 3b4907c23cd3972b6113231ac303f293ad28d6b5