Introduce SLA Violation Controller
Summary:
This diff introduce SLAViolationController.

It works by measuring the EWMA (exponential average over a sliding window) of the latency violation of the reponses (in other terms, the percentage of response that exceed the SLA of the server), then compare that ratio to a predefined threshold.

If the ratio is above the threshold, it starts rejecting incomming request in a linear way, starting at 0% at the threshold.

Reviewed By: yfeldblum

Differential Revision: D13554698

fbshipit-source-id: 2866d449c4fff6c9da34a2e3912a85490ae3f080