Use a constant to indicate an unknown amount of data remaining
Summary:
The low-level protocols don't have any way of knowing how
long the message is without going through and parsing it.  Use a
constant instead of magic number.

Drive-by fix to RemainingBytes() for zlib transport.  The old
implementation would fail when used with the binary or compact profile
and binary or string fields.

Reviewed By: fiorix

Differential Revision: D14230495

fbshipit-source-id: a3dd81f98184b348ef2f2e20de051cbdbf669777