Fix unittest flakiness related to stack destructor race.
Summary:
This fixes a stack destructor race that I introduced on Friday with the
request checksumming change.

Particularly, the Batons used to pause the EventBase so that we could "corrupt"
the data for checksumming need to be heap allocated as shared_ptr, rather than
stack based. It's not really a good idea to have another thread have a pointer to
another thread's stack, even if we expect it to be done.

Reviewed By: vitaut

Differential Revision: D14407522

fbshipit-source-id: dcd5f85562f9813249d0629647b53e7ab1c72936