Add Folly to the compiler_lib CMake link
Summary:
In D14327209 compiler_lib became dependent on Folly, but there is no matching change to
introduce this dependency to the CMake (open source) builds.
For the majority of CMake compile configurations we "get aways with it",
because folly has been installed to the system include path, but in the case of
LogDevice, we build dependencies in submodules, without installing Folly to the
host system; hence the build fails.

Reviewed By: avalonalex

Differential Revision: D14439045

fbshipit-source-id: dddc6429be6a0669bd6807a67abf9b579c11f505