Java: THeaderTransport `getBytesRemainingInBuffer` fix in case of deprecated unframed transport
Summary:
`TheaderTransport` is a subclass of `TFramedTransport` which read data on a per-frame basis.

Unfortunately, in Java, THeaderTransport is also used with the deprecated unframed transport. Then the `getBytesRemainingInBuffer` method doesn't really make any sense, since it returns the number of bytes in the buffer and not in the frame.

This change change the behavior of `getBytesRemainingInBuffer` in the case it is used with an unframed transport. In that case it just returns `-1` which indicates the transport is unbuffered.

Reviewed By: yfeldblum

Differential Revision: D14510874

fbshipit-source-id: 57434153d98032ff9678ee56def7213e7bacd7f7