Extend types fixture with annotated types
Summary:
Introduce a new struct for testing annotations on types.

The new struct introduces two fields, one of them is annotated with `noop_annotation`. Annotating the second one would hit compiler bug and result in unresolved type, which would be surfaced to the user as if the type is not defined.

Stacked diff fixes the bug and adds annotation to the second field.

Reviewed By: yfeldblum

Differential Revision: D14529328

fbshipit-source-id: 49692c1f468a3f40a9cfac349569105c4eb9b4e4