support 'message' annotation for thrift exceptions, to control str(exc) behavior
Summary:
str(exception) is the defacto python way to access the exception message from an Exception instance.
Right now it defaults to the call args of the Exception, as a tuple, or just the first argument if there is only one.  This is default from Python.

In thrift-cpp2 and thrift-py both support the "message" annotation on exceptions to control the what()/str() respectively,  So this adds the same support to thrift-py3

Reviewed By: danshiovitz

Differential Revision: D14622368

fbshipit-source-id: b73655c8ed12e400e7e2ea75005a5c8a55c406dd