make is_string, is_binary and is_string_or_binary more clean in t_base_type
Summary:
currently, `is_string()` in  `t_base_type` can return True is the base type is either `string` or `binary`. with this diff, we introduced a new method `is_string_or_binary()` which retain the same behavior as `is_string()` before this diff to make it more clear. We then make `is_string()` to return True if and only if when base type is actually a string and `is_binary()` to return True if and only when base type is actually a binary.

We then update the usage of `is_string()` to `is_string_or_binary()` whenever makes sense.

Reviewed By: stevegury

Differential Revision: D14678152

fbshipit-source-id: 2bfda77e7035d5becdaf1a873c018b5dec14e852