remove duplicated enum value from generated Enum.fromInteger(...)
Summary:
When `thrift.duplicate_values` is enabled, it is possible for multiple `t_enum_value` with different `name` to share the same `value`, leads to error in `<Enum>.fromInteger`.

In this diff, we remove duplicated enum value from generated `Enum.fromInteger(...)` by deduplicating the integers in the switch and then, for any integer with multiple associated constants, to choose the first of those constants in IDL order.

Reviewed By: yfeldblum

Differential Revision: D14678156

fbshipit-source-id: 9ac72d1b063ca67c7b0ea2f5893101d8f4d87355