Python: Keep trailing '.' for float/double constants
Summary:
Python coerces any constant without trailing '.' as a integer.
In the case where we have float/double fields with integer default values, this will change the type of the field.

This diff changes the way we display constants, in the case of float/double we ensure we add trailing '.'.

Reviewed By: vitaut

Differential Revision: D14809508

fbshipit-source-id: ca2de9ff8bba642c00ab00229e6a0786a75591ea