Remove most remaining RequestRpcMetadata allocs
Summary:
With the previous Rpc Metadata alloc removals in now, it looks possible to do a final
round of removing heap alloced RequestRpcMetadata.

The one case that I din't convert was (non-rocket) ThriftClient, which hops a thread
boundary with the metadata, and hence it's not completely obvious that struct-wide
std::move() would be better than a moved alloc.

Reviewed By: jmswen

Differential Revision: D14877055

fbshipit-source-id: fc5232451d28ca1c1e3e611325cc1ff6681d693a