Installation Steps for thrift-py3
Summary:
To ensure consistency between Cython extensions and c libraries we build thrift-py3 extensions using CMake PythonLibs. CMake PythonLibs includes no installation logic. In this change we use setuptools to install the extension; overriding the build_ext command with a command to copy the library we built during compile/link phase.
This method of overriding python setuptools  was inspired by reviewing the cython-distutils and scikit-build; both use a similar method to supplant the build of the extension with their own logic.

Reviewed By: vitaut

Differential Revision: D14832844

fbshipit-source-id: 6a508bc4bc9f71c7f6f75607cb0cc22cae405fbf