move deduped_values() from t_mstch_swift_generator to t_mstch_objects
Summary:
When trying to move cpp reflection generation logic from compiler/py to mstch_cpp2_generator, we encountered the same problem due to `thrift.duplicate_values`. In this diff, we move `deduped_values()` method to t_mstch_objects as it will be useful in the following diffs.

In the long run, we need to get rid of `thrift.duplicate_values` and migrage usage of `enum:dedupedValues` to `enum:values`

Reviewed By: stevegury

Differential Revision: D15033483

fbshipit-source-id: b1c48c8bb1dd6bc936486fb123d769549ef7fe99