Avoid allocing/copying names for ContextStack.
Summary:
We were previously copying the serviceName/method string names into
the ContextStack for every method call, which was a little unfortunate
because they were largely compile-time constants.

We can't just trivially unconditionally keep the passed in const char*,
since the interface is broad enough for some users to pass in a
temporary string c_str() to it (this happens).

That said, in the code we generate, we know when we're passing in a
compiler-time constant string. So we can just add an alterate calling
codepath, where we  wrap the const char* in a trivial struct to signal
that it's legal for ContextStack not to copy.

Reviewed By: vitaut

Differential Revision: D15047761

fbshipit-source-id: 011f2c95e613f22fc389ea2e099d24ea5ba92d6e