Use shapes for IThriftStruct::SPEC
Summary:
In order to have some structured, typed access to the fields in `SPEC`, we define `TFieldSpec` (and `TElemSpec` for collections, which is `TFieldSpec` minus `var` and `union`) which lists all the possible fields and their types.

Hack doesn't yet support recursive shapes, so unfortunately the best we can do for nested collections is to type `elem`, `key`, and `val` as mixed. There's no subtyping of shapes so we have to spell both shape types out entirely, despite one being a superset of the other.

Reviewed By: yfeldblum

Differential Revision: D15074782

fbshipit-source-id: f23009abbd471daeed933983deb545c55d14197e