Remove NameWrapper from ContextStack optimization.
Summary:
Previously, we checked in code to avoid copying const char* to std::string
in ContextStack on a per-RPC basis.

But to do so, we added a wrapper, NameWrapper, to allow both passing the
const char* when it was whitelisted as safe (generated code), as well as
copying to a string when not whitelisted.

That said, since this is an internal api, it makes most sense and is far
simpler simply to require that the lifetime of the const char* to last
for the duration of the RPC. In all the thrift-synthesized cases, the data
lasts for either the length of the service or is compiler-constant.

Reviewed By: yfeldblum

Differential Revision: D15112352

fbshipit-source-id: 70231632e9a1027a35cf67f30dcf315022ee62ff