(Thrift/Streaming) max batch process size in async StreamGenerator
Summary: The previous StreamGenerator for future has a flaw. The number of concurrent Futures issued bonded to the credit requested from client. In the extreme case if client send request with kNoFlowControl which equals to int64_max will trigger creating this many number of futures and hang the server forever. This comes down to when creating stream, server side owner need to specify a max number of futures that can be issued out concurrently. For coroutine async generator use case this number will always be one.

Reviewed By: lewissbaker

Differential Revision: D15036582

fbshipit-source-id: e1476fe26fd7fe9777a0343ff18bb399f44834e5