Relax size requirements for Frozen tests
Summary: `std::pair<uint32_t, uint32_t>` is now trivially copyable, which triggers the conservative criteria Frozen2 uses for aligning storage. This should be fixed by modelling alignment in the layout types, but in the meantime this will lead to ranges of ranges being filled with significant padding bytes. I'll follow up with a proper alighnment fix, but for now I'm just relaxing the tests.

Reviewed By: yfeldblum

Differential Revision: D15231572

fbshipit-source-id: baf97409e07749855a6a7e59fab11d7f8e030e94