(Thrift/Streaming) Simplify StreamGenerator for coro::AsyncGenerator
Summary:
Previously coro::AsyncGenerator was supported by converting AsyncGenerator to a bunch of coro::Task then to a semiFuture, and integrated as generator that produce semiFuture<T>. Taken the idea of lewissbaker from comment in D15036582, It is kind of waste of doing this as there is nothing prevent directly supporting AsyncGenerator with coroutine, by doing this we also get rid of yarpl in this piece of logic.

Preemptive cancel support will be added in a following diff.

Reviewed By: lewissbaker

Differential Revision: D15106223

fbshipit-source-id: de31761e08c19f84988fc8c29efd021db4264161