Create DebugString util, to print serialized protocol data
Summary:
This adds a DebugString util library to take serialized protocol data (e.g.
CompactProtocol), and to produce a human-readable output from that,
without needing the Thrift struct definition.

We won't be able to include the tag names based on the Compact/Binary Protocol
data, but we still provide some useful debugging info. The use case I have in
mind is rendering arbitrary Thrift-encoded data in the row of a random table
somewhere, or that is the result of some computation.

An example of the output from this module would be like the following:

toDebugString() ->
  struct {
    1: string = "image.jpg"
    2: i32 = 640
    3: i32 = 480
    4: float = 0.85
    5: list<i64> = [
      2342342343
    ]
  }

We also provide a reverse fromDebugString() parsing function, to take the
string representation above, and convert it to a CompactProtocol/
BinaryProtocol/etc. stream.

We provide a few simple rendering options, such as whether newlines
are wanted and indent level. Performance could probably be better, but
can optimize that if needed.

Reviewed By: vitaut

Differential Revision: D15162411

fbshipit-source-id: c9043c471241bf52cc900d3236bff69a5e6ba763