(Thrift/Streaming) Support CancellationToken-based canellation in AsyncGenerator stream
Summary:
The `toAsyncGenerator()` function now takes an optional `CancellationToken` parameter that allows the consumer to cancel a stream when a consumer is currently suspended waiting for the next value.

Adds a guard in EmitterSubscription: to handle the case where a call to onError()/onValue()/onComplete() comes in after the stream has been cancelled and the inner subscription has been released.

Reviewed By: rhodo

Differential Revision: D14893470

fbshipit-source-id: 30646bf1e96273a550dab1e78a94a4cf6bf6a3d6