Make onDetachable logic easier to integrate with low-level stream API
Summary:
Currently, the stream flowable has some logic to potentially call the `onDetachable()` callback
when the stream terminates. In order to simplify the new low-level streaming API (i.e., not having to worry about
explicitly calling `onDetachable()` in any low-level callback), we can push most of this logic down to the `RocketClient`
stack, which already has a couple hooks in the right spots to possibly invoke `onDetachable()`.

Reviewed By: yfeldblum

Differential Revision: D15275455

fbshipit-source-id: bd0a78ba0f5069af9628fd6bc0069266001c32c2