In Rocket client, convert some throw paths to Try<>
Summary:
This change converts some 'throw exception' cases in the Rocket client side
to wrap the exception and return, like we do on the Rocket server side.

The main motivation for this is re: routing-type clients and preventing
cascading failures that might related to high exception processing costs.
For instance, a burst of timeouts on the client side, which due to high
cpu cost, will lead to further timeouts.

Reviewed By: yfeldblum

Differential Revision: D15342510

fbshipit-source-id: b2367a9ab1baddcf38b3d98ac403232375815862