(Thrift) benchmarking Thrift and ProtoBuf on some more complex structure
Summary: ProtoBuf version 3.6.1 and proto3 syntax was used in this experiment. Apart from the benchmark test already there in the repo, two kinds of new benchmarks are added. One is a large struct with many fields. The other is a nested map, since Protobuf doesn't support map to be the value a map (no map<key, map<key, value>>) and need to add a wrapper message in between, I did the same in Thrift to make it fair.

Reviewed By: vitaut

Differential Revision: D9381822

fbshipit-source-id: 9d9b08bc14aecbfe2b5e103a716697781fa04845