declare has_errors global
Summary: In python, it seems global scope is a read-only shadow within the local function/method scope: you can access it, providing there’s nothing declared with the same name in local scope, but you can’t mutate it. In order to mutate global scope variable, we need to declare it global.

Reviewed By: yfeldblum

Differential Revision: D15488396

fbshipit-source-id: 7f3ff9fae07cd5d26890767768c4026d5cab6203