Don't copy shared_ptr<TServerObserver> multiple times per request
Summary:
Each `ThriftServer` has one `TServerObserver`. A `shared_ptr<TServerObserver>` is copied and destroyed
multiple times per request; the shared count is highly contended.

This diff optimizes `getObserver()` so that there is minimal contention on the hot path. While we're at it, we solve the inherent raciness of the `getObserver()`/`setObserver()` API by making it so that `setObserver()` can only be called once per server (which matches practical usage of the API).

Reviewed By: spalamarchuk

Differential Revision: D15476889

fbshipit-source-id: 39ee00da1038c9b440ba80c3cfc8865ece8ccd99