introduce java.swift specific annotation java.swift.annotation
Summary: In this diff, we introduced a new java.swift specific thrift annotation `java.swift.annotation = <special annotation>` which allows java swift generator to annotate generated java classes and fields with a custom Java annotation. This is useful when thrift users like to use some annotation based processing on generated thrift classes. This feature can be dangerous and we consider this to be an alpha, and may remove once we have a better alternative.

Reviewed By: yfeldblum

Differential Revision: D15527680

fbshipit-source-id: 2b281c9b4f3a638b780e37619104fdab5b510d78