Single allocation response serialization
Summary:
Previously, we'd do several allocations for every response + potentially copy the serialized response (one for the response body, another for transport framing and/or metadata).

With this change, we reserve 128b of headroom in the response buffer and use it for serializing metadata and frame header in the rocket transport (any other transport can do the same).

Reviewed By: yfeldblum

Differential Revision: D15712649

fbshipit-source-id: c1cd5bf24756292554886476a5e2fdac8b3ba265