Disable check for tautological constant compare
Summary:
Newer clang compilers check for tautological constant comparisons.

These ones are a bit weird because of enums - variables are allowed to take on values which do not correspond to any specific enumerator in the enum, although compilers sometimes like to assume they can't. So cast to the underlying type to evade this assumption, keep the runtime checks in the code, and keep the warning enabled in the compiler.

Reviewed By: yfeldblum

Differential Revision: D15771955

fbshipit-source-id: 0ea33f8e7acca03a5f98fa2764767d266be86830