Don't access channel after calling it in generated sync_ methods
Summary:
Previously, we had `sendRequestSync` that would block and upon return we wouldn't use the client or the channel in the `sync_` methods.

Recent changes changed that and now we access `channel_->getEventBase()` after sending request.

With rocket transport being used on fiber, this would mean that we already got a response and potentially the `channel_` is not valid. This changes the code to capture `eventBase_` before calling into channel.

Reviewed By: yfeldblum

Differential Revision: D15818782

fbshipit-source-id: c36da9e8d8faeed8ad94b3cb67c70145bbf6b67e