Use low-level streaming API in RocketServerConnection
Summary:
This diff reimplements the lowest-level streaming functionality in the `rocket` server stack. `RocketServerConnection` now maintains a pointer to a `StreamClientCallback` (new low-level streaming API) instead of a `yarpl::flowable::Subscriber`.

This diff is the first (and most major) part of supporting the low-level streaming API in `AsyncProcessor`, similar to what already exists on the client side in `RequestChannel::sendRequestStream()`.

This diff is also a step towards optimizing the memory footprint of a stream on the server side. (The idea is to replace the `yarpl` parts with a few memory-optimized components, including `RocketStreamClientCallback` and `StreamServerCallbackAdaptor` introduced in this diff. Note that `CoroStreamImpl`, which already exists, is also a step in this direction, higher up the server-side streaming stack.)

Reviewed By: rhodo

Differential Revision: D15841387

fbshipit-source-id: 2fcce2f47c752eb3dfee2e49e53d48980bd49b28