Extend types fixture with structs that require topological sorting in C++ codegen
Summary:
This extends a types fixture with a case when we use a struct before declaring it. The purpose of it is to test that C++ codegen would perform appropriate reordering when generating structs.

Note, currently we don't do this reordering, it's done in the stacked diff.

(Note: this ignores all push blocking failures!)

Reviewed By: stevegury

Differential Revision: D12965408

fbshipit-source-id: 92e13a1cae7597b9b0cc735d4b5657e7bad9dde3