Remove enum check + generation of IntRangeSet
Summary:
Remove the creation of the `VALID_VALUES` set for enum validation.
This posed problem with extremly large enum entries and it is not very useful.

The behavior is now similar in C++, an invalid enum entry will be map to `null`
instead of forcing the thrift struct to be rejected. This provide a better
forward compatibilty.

Differential Revision: D15814168

fbshipit-source-id: a57527db66bf124b9315f1ae3fc333e6850d6835