Thrift server: allow service handler to provide a default folly::RequestContext
Summary:
For library that supports async binary, through coroutine and/or thread. A common pattern is to provide global config through folly RC.
This allows end user to choose the scope of the feature usage, global (by initializing before the thread) or per-thread (after initialization), per-request (after creation of the root coroutine)...

Currently, thrift server doesn't allow global or per-thread initialization, as it always reset folly RC, before processing a request.
This diff provides a way for user to provide a base RC to be copied during pre-processing.

Reviewed By: andriigrynenko

Differential Revision: D15958090

fbshipit-source-id: 01c338f65b85c4b1b50b539d4b33b4dcae11e2e0