Add max_length to randomizer constraints for collections
Summary: Fuzzing thrift is really useful however we shouldn't test the degenerate cases where random decides to pick 1 short of int max for a string length. So this diff adds the ability to set a constraint that will limit the length returned for _get_length for collections (set, list, map, string).

Reviewed By: yfeldblum

Differential Revision: D15943369

fbshipit-source-id: 8b15a51dc95f8e5cd89312f9666c947afa98cd7d