Deprecate enable_haddock
Summary:
`enable_haddock` used to be a legacy way to hide TARGETS away from the haddock build if we don't want to include them in the haddock documentation. However, currently it's only in use in the Haskell Thrift Compiler to control whether the Haddock Markup should be generated for the Haskell Code or not. This is not very useful, there's no problem in adding the haddock markup to all generated modules.

Note: this only impacts Haskell Targets and the Haskell Thrift Compiler

Reviewed By: zilberstein

Differential Revision: D15923866

fbshipit-source-id: ac37092d91764244f35abc9f2f5e12e0170d2c4c