Remove logic that allows waiting on first payload for stream request
Summary:
This diff reverts D15307579, with some tiny non-logic changes. The idea of D15307579 was to allow the user
to await on the first response on a stream, but it would not actually be safe to do this in the current implementation
since it is possible for incoming frames on the stream to be silently dropped. To make sure no one tries to use this
functionality, let's back out the change.

Reviewed By: yfeldblum

Differential Revision: D15928014

fbshipit-source-id: 005977410ccfb6530b0ef6f26a306ae51ffc24a2