Destroy request inline when already on EventBase, try 2
Summary:
There's no need to postpone destruction of the context and we can do this inline. This significantly improves performance of high-throughput services with inline processing (like Memcache) due to better reuse of the memory (with jemalloc, we simply keep reusing the same object over and over again).

The only side-effect is that now active_requests becomes real-time, because the decrement is done in the destructor of the context instead of being accumulated for the duration of event loop.

Reviewed By: jmswen

Differential Revision: D15712654

fbshipit-source-id: e1507f351cb945ea49df567668f5ee9df1534a09