Handle DNS Resolution Completely from Python, make ssl clients full featured
Summary:
We had a back and forth with C++ before, we created a promise and had c++ await on its future to get a resolved hostname.

This time, if we think it could be a hostname we use ipaddress to check if its an address, if it fails then we return a wrapping async context manager that will handle the resolve, and wrap a standard get_client

While I was removing all the promise garbage, I noticed SSL clients didn't properly handle client types and protos so I added that logic, to make SSL sockets exactly like non ssl clients in reguards to how it handles http/protocal/and client type.

This has the added benefit of less log spew during cancellation.

Reviewed By: sdunster

Differential Revision: D15142815

fbshipit-source-id: 54923cda0078615aa36ee69c1e4d835486a962b8