Synchronize Hack files
Summary:
The hack files were out of sync, the source of truth for those files are
outside of this repository.

I ran the `ImportThriftLibFromWWW` script which only works for Hack files, I
updated it to remove the transpilation step which was supposed to update the
php files (it doesn't work anymore since it can't find the `h2tp` transpiler)
and AFAIK we don't use php anymore.

I also ran `arc lint` to update the license headers to the correct format.

Differential Revision: D16186893

fbshipit-source-id: 1ceafbeb43bfdcb63315b615212680811b243fed