Make object_partial() return
Summary: `object_partial()` is templated to have a return type of `folly::Function<Ret(Args...)>`. However, the current implementation only supports `Ret = void` because the inner lambda doesn't return the result of the C-style function-pointer call. In the current Thrift py3 implementation only `void` returns are used so this has not yet been a problem.

Reviewed By: yfeldblum

Differential Revision: D16133347

fbshipit-source-id: a9805cbe6b873e17efeba08adf608f0c1a1e7237