Fixtures refactoring
Summary:
The basic fixtures does more than the basic thrift usage, it is now a
repository for plenty of the annotations that we support.
In order to clarify the basic usage and simplify the job of the reviewers of
the next commits, I split the `basic` fixtures into `basic` which only contains
basic features (struct, enum, union, services) and `basic-annotations` which
contains the differents annotations we support.

I also added a `union` in basic, which was lacking.

Reviewed By: yfeldblum

Differential Revision: D16090433

fbshipit-source-id: 3545fb34bc68fe03160941fd2bf27ee6d5bd1ada