Move iobuf Python wrapper from thrift to folly
Summary:
We expose iobuf to Python via the folly.iobuf Python extension. The code for this extension has historically been in the fbthrift source tree. Move it to folly so that folly produces folly extensions and thrift produces thrift extensions.

Without this change we end up with the folly.iobuf Cython extension being packaged with thrift rather than folly, leading the need for a number of work-arounds to allow Python and Cython to find folly.iobuf without stopping their searches on folly.executor.

The tests for IOBuf python extension has been added to folly while maintaining the thrift-specific IOBuf tests in thrift.

Reviewed By: vitaut

Differential Revision: D14405645

fbshipit-source-id: 662a4e32c219f6d044bb6e903a2525ec1b4059c5