Fix Thift randomizer seeds for Structs
Summary: Thrift randomizer can take in seeds. These seeds are usually the same type as the type being generated. However to generate structs the randomizer needs to be able to treat the seeds as a dict. This works for the most simple structs that have no fields that are themselves other structs. However if things get more complex then it becomes very unweildy to deal with dicts. So this diff makes it possible to use a struct with the thrift randomizer seeds.

Reviewed By: paulmar

Differential Revision: D16045032

fbshipit-source-id: e11b578554c90b5649a6e8ae8bb22a4fc976fad2