Fix race between disconnect and onFirstResponse
Summary:
This diff involves the following race:
* Client sends a streaming request to server.
* Server starts processing the request.
* Client disconnects.
* Server handler completes and `onFirstResponse()` is called.

In this scenario, `RocketServerConnection` and the stream will be leaked -- `RocketServerConnection` will be in the `CLOSING` state, but the actual cleanup will never happen.
Also, the stream will never be canceled or have a chance to make progress since no further credits are incoming. This is fixed by checking in `onFirstResponse` whether we need to
trigger connection shutdown via `RocketServerConnection::closeIfNeeded`.

Reviewed By: rhodo

Differential Revision: D16368319

fbshipit-source-id: 8e0d23d1f4bd72dd028bc0051ace84cad6065de5