Generate variable names prefixed with "__" to avoid conflict
Summary:
The `javadeprecated` codegen generate code with variables named `field` and
`value` which are not considered as keyword in Thrift. So it's possible to have
a Thrift struct/union using these names.
The generated code is therefore invalid and doesn't compile.
This diff fixes the problem by prefixing the generated variables by `__` to
avoid naming collision.

Reviewed By: yfeldblum

Differential Revision: D16295942

fbshipit-source-id: 66b7d7856fa52bb9e7535b4a4b128788011c966f