Java: Remove generation of VALUES_TO_NAMES in enum
Summary:
Now that we are generating proper Java Enums, there is no real need for a public `VALUES_TO_NAMES` map.

Java enums provides a `values()` method which returns a list of values, you can also call `.name()` or `.getValue()` to respectively access the string representation or value of the enum.

As of this diff, `VALUES_TO_NAMES` is not used anywhere outside thrift.
It was only used in the `toString()` method to display the string representation of the enum, this can now be replaced by `myEnum.name()`.

The `findByValue(int value)` find an enum based on its value, it is currently doing a linear scan of all the entries.
I improve the implementation so that when the number of enum is small (< 16) it still does the scan, but if it's big, it relies on a statically initialized map still called `VALUES_TO_NAMES` but private this time.

Note: `IdentityHashMap` is better suited (more efficient) when the key is a primitive.

Reviewed By: vitaut

Differential Revision: D15916681

fbshipit-source-id: 0341e9f740f486994cc239f12d1d6cb228e9bbbd