move all thrift services to use context based server implementation
Summary:
This moves ALL go projects to use Context based thrift servers.

behold

Files found using this:

```
$ for file in `fbgs --max-bytes 50000000 -f TARGETS go_thrift_namespaces | cut -d: -f1 | sort -u`; do grep -q use_context $file || echo $file; done | sed 's!fbsource/fbcode/!!' > ~/TODO
```
then used codemod:

```
$ codemod -m --file-list `cat ~/TODO | paste -d, -s` 'thrift_library\(' 'thrift_library(thrift_go_options="use_context",'
```

That was not enough though, I had to `jf submit` here to se what was failing and iterate.

Your team's review needed for D15941043. Move all thrift server to context based implementation.

If you own a golang service in fbcode please take the time to review this diff.

As much as I trust unit tests, and cogwheel tests and other sandcastle runs, and the fact that ours is a statically typed language (which makes refactoring easier) I still would like people to check that out, especially teams that are heavily golang based.

## What is this?
One of the things Go Foundation has in its roadmap is moving all thrift servers to context based implementation.
What does this give us? Various things, some of which are:

* The opportunity to enrich our thrift handler with interceptors. An interceptor is a function that runs before the actual thrift method that was called and is able to view and/or modify both the arguments and the results - as they are returned to the caller - of that method
* The interceptor implementation then gives us the opportunity to enrich our thrift handlers with useful information from the TLS layer, most notably the TLS client identity.
* This finally will give us the opportunity to fully integrate all of our thrift servers with hipster/aclchecker.

This is done by:
* enabling `thrift_go_options="use_context"` in all thrift_library
* Move servers to `thriftbase.ServerContext`
* Other bits and pieces, since enabling context changes the service interface this was breaking the client implementation (which doesn't have context). Lucas kindly offered to work on D16377914 which implement a specific <Service>ClientInterface. The diff moves projects to that, also moving gomocks to those interfaces.

## A 6K LOC diff? you crazy angelo? y u no hg split?

I thought about this, and I decided to go for a single diff because it's easy to unland it in case of emergency.
Also the way things works with buck's TARGETS files, once you enable `thrift_go_options="use_context"` in a TARGETS file that triggers a domino effect where you need to enable that on everything...
I think the best way to review this (given the size of the diff), is to jf get it and try compile/test your team's code.
I am pretty confident things work, things are building, unit tests are green, I tested the provisioning pxeboot stack and it works.

Reviewed By: mkaczanowski

Differential Revision: D15941043

fbshipit-source-id: c4b1897f10990399e825d9d6d4687bbaa8ff97df