Add capability to set flush list from application
Summary:
Currently, we postpone flushing pending writes to the socket until the end of next event loop. In certain applications, such as routing libraries, we may want to postpone this even further to achieve better batching.

This diff exposes API for application to attach centralized flush list and control when flush is happening.

Reviewed By: jmswen

Differential Revision: D16626088

fbshipit-source-id: 772f52f3773078c888a5fa98c9e5da2e1dba2a6b