Clean up the chatroom client example
Summary:
Make the chatroom client example not depend on `thrift/perf` because the latter is
an internal benchmark and example is supposed to demonstrate how to use the
public Thrift API. Also simplify it by removing the transport flag hardcoding
the rocket transport.

This also resolves a problem when building this example with CMake (due to missing dependency on the perf target).

Reviewed By: yfeldblum

Differential Revision: D16929543

fbshipit-source-id: bd35b7e94c5fda6266f6281803a1eb3b63b4601a