Let protocols pick their own choice of types.
Summary:
Some protocols (e.g. Nimble) don't represent the full thrift type on the wire.
This change makes that fact explicit. Otherwise, we'd have to make all of our
type comparisons (e.g. in detecting mismatches) go through a layer of
indirection.

This is not a morally complete change; VirtualProtocol needs an interface
change to allow it to forward any type of enum to the underlying
implementation, and so take an integer of some sort (this compiles only because
we don't yet test VirtualProtocol and Nimble together). However, I'll leave the
change for a subsequent diff.

Reviewed By: bithree

Differential Revision: D16930579

fbshipit-source-id: dd46e3090ca08d1f3fdd58b7c3b00f8853f12a52