break dependency on python-future
Summary:
python-future was only used for a single iteritems call, which is
trivially replaced with items(). In Python 3, this change is a
no-op. In Python 2, since the iterator must be converted into an
argument list here, it may even use less memory. Either way, .items()
is faster than .iteritems() in Python 2, and the difference is
unlikely to matter.

Reducing an external dependency makes packaging easier, especially on
Windows.

Reviewed By: yfeldblum

Differential Revision: D16997250

fbshipit-source-id: 32f9f9c6e32caa7df9793f883fcbcda4b69cfc41